+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - After token generation direct user to the settings page.
 - Moved generation of token file into web interface.
+- Update logging to go to file and StdOut.
 - Update release to remove all .go files from subdirectories.
 
 ## [v0.1.0]